# Conflicts:
#	apl-core/src/main/java/com/apollocurrency/aplwallet/apl/core/app/FundingMonitor.java
#	apl-core/src/main/java/com/apollocurrency/aplwallet/apl/core/dao/appdata/UnconfirmedTransactionTable.java
#	apl-core/src/main/java/com/apollocurrency/aplwallet/apl/core/entity/blockchain/UnconfirmedTransaction.java
#	apl-core/src/main/java/com/apollocurrency/aplwallet/apl/core/entity/state/currency/CurrencyType.java
#	apl-core/src/main/java/com/apollocurrency/aplwallet/apl/core/http/get/GetConstants.java
#	apl-core/src/main/java/com/apollocurrency/aplwallet/apl/core/transaction/ShufflingTransaction.java
#	apl-core/src/main/java/com/apollocurrency/aplwallet/apl/core/transaction/messages/AbstractAppendix.java
#	apl-core/src/main/java/com/apollocurrency/aplwallet/apl/core/transaction/types/ms/MSCurrencyDeletionTransactionType.java
#	apl-core/src/main/java/com/apollocurrency/aplwallet/apl/core/transaction/types/ms/MSCurrencyMintingTransactionType.java
#	apl-core/src/main/java/com/apollocurrency/aplwallet/apl/core/transaction/types/ms/MSCurrencyTransferTransactionType.java
#	apl-core/src/main/java/com/apollocurrency/aplwallet/apl/core/transaction/types/ms/MSPublishExchangeOfferTransactiionType.java
#	apl-core/src/main/java/com/apollocurrency/aplwallet/apl/core/transaction/types/ms/MSReserveClaimTransactionType.java
#	apl-core/src/main/java/com/apollocurrency/aplwallet/apl/core/transaction/types/ms/MSReserveIncreaseTransactionType.java
#	apl-core/src/main/java/com/apollocurrency/aplwallet/apl/core/transaction/types/ms/MonetarySystemExchangeTransactionType.java

+++ apl-core/src/test/java/com/apollocurrency/aplwallet/apl/core/migrator/ReferencedTransactionMigratorTest.java
@@ -88,7 +88,7 @@ class ReferencedTransactionMigratorTest {
             stmt.executeUpdate("update referenced_transaction set height = -1 where db_id >= 40");
         }
         List<ReferencedTransaction> notMigrated = CollectionUtil.toList(referencedTransactionDao.getManyBy(new DbClause.LongClause("db_id", DbClause.Op.GTE, 40), 0, Integer.MAX_VALUE));
-        notMigrated.forEach(rtx -> assertEquals(-1, rtx.getHeight().intValue()));
+        notMigrated.forEach(rtx -> assertEquals(-1, rtx.getHeight()));
 
         migrator.migrate();
